requirements.txt
alembic
black
fastapi
pydantic-settings
sqlalchemy[postgresql_asyncpg]
uvicorn[standard]

